+++ daiv/codebase/managers/issue_addressor.py
@@ -2,7 +2,6 @@ from typing import cast
 
 from django.conf import settings
 
-from langchain_community.callbacks import get_openai_callback
 from langchain_core.prompts.string import jinja2_formatter
 from langchain_core.runnables import RunnableConfig
 from langgraph.checkpoint.postgres import PostgresSaver
@@ -63,7 +62,7 @@ class IssueAddressorManager(BaseManager):
 
         config = RunnableConfig(configurable={"thread_id": generate_uuid(f"{self.repo_id}{issue.iid}")})
 
-        with PostgresSaver.from_conn_string(settings.DB_URI) as checkpointer, get_openai_callback() as usage_handler:
+        with PostgresSaver.from_conn_string(settings.DB_URI) as checkpointer:
             issue_addressor = IssueAddressorAgent(
                 self.client,
                 project_id=self.repository.pk,
@@ -71,7 +70,6 @@ class IssueAddressorManager(BaseManager):
                 source_ref=self.ref,
                 issue_id=cast(int, issue.iid),
                 checkpointer=checkpointer,
-                usage_handler=usage_handler,
             )
             issue_addressor_agent = issue_addressor.agent