+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Removed unused `get_openai_callback` on codebase managers.
+
 ## [0.1.0-alpha.8] - 2024-12-11
 
 ### Added